refactor: organize OIDC authentication helpers
