+++ src/controllers/users/update-password.ts
@@ -1,9 +1,9 @@
 import { and, eq } from 'drizzle-orm'
 import { getContext } from 'hono/context-storage'
 import { HTTPException } from 'hono/http-exception'
-import { getAuthMode } from '#@/constants/auth.ts'
 import { sessionTable, statusEnum, userTable } from '#@/databases/schema.ts'
 import { hash, verify } from '#@/utils/server/argon2.ts'
+import { getAuthMode } from '#@/utils/server/auth.ts'
 
 export async function updatePassword(currentPassword: string, newPassword: string) {
   if (getAuthMode() !== 'local') {
