+++ src/server/node-dev.ts
@@ -1,6 +1,6 @@
 import { Hono } from 'hono'
 import { RouterContextProvider, createRequestHandler } from 'react-router'
-import { getAuthMode } from '#@/constants/auth.ts'
+import { getAuthMode } from '#@/utils/server/auth.ts'
 import app from './app.ts'
 
 getAuthMode()
